# modules/static/swagger.yaml
swagger: "2.0"
info:
  title: "쫑이까까후원재단 API"
  version: "1.0.0"
  description: "멸종위기종 보호를 위한 REST API 문서"
basePath: "/"
schemes:
  - "http"
  - "https"
...
paths:
...
  /inspection/normal/search:
    get:
      tags:
        - Search
      summary: 일반 검수 이미지 검색
      security:
        - Bearer: []
      parameters:
        - in: query
          name: project_name
          type: string
          description: 프로젝트 이름
        - in: query
          name: date
          type: string
          description: 검색할 날짜 (YYYY-MM-DD)
        - in: query
          name: serial_number
          type: string
          description: 카메라 시리얼 번호
        - in: query
          name: species
          type: string
          description: 동물 종
        - in: query
          name: page
          type: integer
          default: 1
        - in: query
          name: per_page
          type: integer
          default: 20
      responses:
        200:
          description: 검색 성공
          schema:
            type: object
            properties:
              message:
                type: string
              data:
                type: object
                properties:
                  images:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        filename:
                          type: string
                        thumbnail:
                          type: string
                        date:
                          type: string
                        serial_number:
                          type: string
                        species:
                          type: string
                        project_name:
                          type: string
                        count:
                          type: integer
                        event_number:
                          type: integer
                        is_classified:
                          type: boolean
                          default: true
              meta:
                type: object
                properties:
                  total:
                    type: integer
                  page:
                    type: integer
                  per_page:
                    type: integer
                  total_pages:
                    type: integer
  /inspection/exception/search:
    get:
      tags:
        - Search
      summary: 예외 검수 이미지 검색
      security:
        - Bearer: []
      parameters:
        - in: query
          name: project_name
          type: string
          description: 프로젝트 이름
        - in: query
          name: date
          type: string
          description: 검색할 날짜 (YYYY-MM-DD)
        - in: query
          name: serial_number
          type: string
          description: 카메라 시리얼 번호
        - in: query
          name: exception_status
          type: string
          enum: [pending, confirmed, rejected]
          description: 예외 상태
        - in: query
          name: page
          type: integer
          default: 1
        - in: query
          name: per_page
          type: integer
          default: 20
      responses:
        200:
          description: 검색 성공
          schema:
            type: object
            properties:
              message:
                type: string
              data:
                type: object
                properties:
                  images:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        filename:
                          type: string
                        thumbnail:
                          type: string
                        date:
                          type: string
                        serial_number:
                          type: string
                        project_name:
                          type: string
                        exception_status:
                          type: string
                          enum: [pending, confirmed, rejected]
                        event_number:
                          type: integer
                        is_classified:
                          type: boolean
                          default: false
              meta:
                type: object
                properties:
                  total:
                    type: integer
                  page:
                    type: integer
                  per_page:
                    type: integer
                  total_pages:
                    type: integer

  # Download 관련 엔드포인트
...
components:
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header
      description: "JWT 토큰을 입력하세요. 예: Bearer {token}"
security:
  - Bearer: []
...